fix oddness in lib/format-json
We had to use setImmediate() before because our call to push() from _read() was
immediately emitting 'readable' on 'this', which caused the consumer to
call _read(), causing _read to be reentrant in a way that we didn't expect.
Each time this happened, we'd leak one logical "outstanding read operation",
which includes a "readable" listener on the upstream json decoder.  We'd get a
warning from Node about a memory leak and the program would quickly slow to a
crawl.